tlv: add parsing Cookie TLV support
This is linked to "Cookie TLV" section from RFC8803:

https://www.rfc-editor.org/rfc/rfc8803.html#name-the-cookie-tlv

  The Cookie TLV (Figure 21) is an optional TLV that is similar to the
  TCP Fast Open Cookie [RFC7413]. A Transport Converter may want to
  verify that a Client can receive the packets that it sends to prevent
  attacks from spoofed addresses. This verification can be done by using
  a Cookie that is bound to, for example, the IP address(es) of the
  Client. This Cookie can be configured on the Client by means that are
  outside of this document or provided by the Transport Converter.

  A Transport Converter that has been configured to use the optional
  Cookie TLV MUST verify the presence of this TLV in the payload of the
  received SYN. If this TLV is present, the Transport Converter MUST
  validate the Cookie by means similar to those in Section 4.1.2 of
  [RFC7413] (i.e., IsCookieValid). If the Cookie is valid, the
  connection establishment procedure can continue. Otherwise, the
  Transport Converter MUST return an Error TLV set to "Not Authorized"
  and close the connection.

  If the received SYN did not contain a Cookie TLV, and cookie
  validation is required, the Transport Converter MAY compute a Cookie
  bound to this Client address. In such case, the Transport Converter
  MUST return an Error TLV set to "Missing Cookie" and the computed
  Cookie and close the connection. The Client will react to this error
  by first issuing a reset to terminate the connection. It also stores
  the received Cookie in its cache and attempts to reestablish a new
  connection to the Transport Converter that includes the Cookie TLV.

  The format of the Cookie TLV is shown in Figure 21.

                        0                   2                   3
    0 1 2 3 4 5 6 7 8 9 0 1 2 3 4 5 6 7 8 9 0 1 2 3 4 5 6 7 8 9 0 1
   +---------------+---------------+-------------------------------+
   |     Type=0x16 |     Length    |             Zero              |
   +---------------+---------------+-------------------------------+
   /                        Opaque  Cookie                         /
   /                              ...                              /
   +---------------------------------------------------------------+

  Figure 21: The Cookie TLV

In this commit, only the parsing part has been added.

+++ convert.h
@@ -94,7 +94,7 @@ struct convert_supported_opts {
 struct convert_cookie {
 	struct convert_tlv	tlv_hdr;
 	uint16_t		reserved;
-	uint32_t		opaque[0];
+	uint8_t			opaque[0];
 } __attribute__((packed));
 
 struct convert_error {
